ref(org-stats): No V1/V2 on frontend
