+++ static/app/routes.tsx
@@ -1275,17 +1275,7 @@ function routes() {
             path="/organizations/:orgId/stats/"
             componentPromise={() =>
               import(
-                /* webpackChunkName: "OrganizationStats" */ 'app/views/organizationStatsV2'
-              )
-            }
-            component={errorHandler(LazyLoad)}
-          />
-
-          <Route
-            path="/organizations/:orgId/statsV1/"
-            componentPromise={() =>
-              import(
-                /* webpackChunkName: "OrganizationStatsV1" */ 'app/views/organizationStats'
+                /* webpackChunkName: "OrganizationStats" */ 'app/views/organizationStats'
               )
             }
             component={errorHandler(LazyLoad)}
