+++ static/app/views/organizationStats/types.tsx
@@ -1,13 +1,28 @@
-export type ProjectTotal = {
-  id: string;
+import {SeriesApi} from 'app/types';
+
+export enum Outcome {
+  ACCEPTED = 'accepted',
+  FILTERED = 'filtered',
+  INVALID = 'invalid',
+  DROPPED = 'dropped',
+  RATE_LIMITED = 'rate_limited',
+}
+
+/**
+ * Raw response from API endpoint
+ */
+export type UsageSeries = SeriesApi & {
+  start: string;
+  end: string;
+};
+
+export type UsageStat = {
+  date: string;
+  total: number;
   accepted: number;
-  blacklisted: number;
-  received: number;
-  rejected: number;
+  filtered: number;
+  dropped: {
+    total: number;
+    other?: number;
+  };
 };
-
-export type OrgTotal = ProjectTotal & {
-  avgRate: number;
-};
-
-export type Point = {x: number; y: number[]};
