+++ tests/js/spec/views/organizationStats/usageChart/utils.spec.jsx
@@ -1,7 +1,7 @@
 import {
   getDateFromMoment,
   getXAxisDates,
-} from 'app/views/organizationStatsV2/usageChart/utils';
+} from 'app/views/organizationStats/usageChart/utils';
 
 const TS_START = 1531094400000; // 2018 July 9, 12am UTC
 const TS_END = 1531180800000; // 2018 July 10, 12am UTC
